separating integration tests, moving postgres setup to helpers
