+++ build.sbt
@@ -1,8 +1,3 @@
-// *** Plugins ***
-
-lazy val root = (project in file(".")).enablePlugins(SbtWeb, PlayScala, DockerPlugin)
-
-
 // *** Coordinates ***
 
 organization := "com.autoscout24"
@@ -83,6 +78,18 @@ version in PB.protobufConfig := "3.0.0"
 PB.runProtoc in PB.protobufConfig := (args => com.github.os72.protocjar.Protoc.runProtoc("-v300" +: args.toArray))
 
 
+// *** Integration tests settings ***
+
+lazy val ItTest = config("it") extend(Test)
+
+def itSpecFilter(name: String): Boolean = name.endsWith("IntegrationSpec")
+def specFilter(name: String): Boolean = name.endsWith("Spec") && !itSpecFilter(name)
+
+testOptions in ItTest := Seq(Tests.Filter(itSpecFilter))
+
+testOptions in Test := Seq(Tests.Filter(specFilter))
+
+
 // *** Test coverage settings ***
 
 coverageMinimum := 80
@@ -97,3 +104,10 @@ coverageExcludedPackages := Seq(
   """toguru\.filters\..*""",         // low test value and hard to test
   """toguru\.app\.ErrorHandler"""    // low test value and hard to test
 ).mkString(";")
+
+// *** Plugins and configs ***
+
+lazy val root = (project in file("."))
+  .enablePlugins(SbtWeb, PlayScala, DockerPlugin)
+  .configs(ItTest)
+  .settings(inConfig(ItTest)(Defaults.testTasks) : _*)
